Goa/*: change unsigned long ip to unsigned int ip per sezero advice for LP64.

+++ Goa/CEngine/gserver.h
@@ -29,7 +29,7 @@ extern "C" {
 
 struct GServerImplementation
 {
-	unsigned long ip;
+	unsigned int ip;
 	unsigned short port;
 	unsigned short ping;
 	HashTable keyvals;
@@ -47,7 +47,7 @@ typedef struct
 } GEnumData;
 
 void ServerFree(void *elem);
-GServer ServerNew(unsigned long ip, unsigned short port);
+GServer ServerNew(unsigned int ip, unsigned short port);
 void ServerParseKeyVals(GServer server, char *keyvals);
 
 #ifdef __cplusplus
